HELP.md updated

+++ src/main/java/com/example/personal_notes_app/Repository/ArticleRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.example.personal_notes_app.Model.Articles;
 
-public interface ArticleRepository extends MongoRepository<Articles, Integer> {
+public interface ArticleRepository extends MongoRepository<Articles, Long> {
 
 }
