+++ src/main/java/com/example/personal_notes_app/Repository/UserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.example.personal_notes_app.Model.Users;
 
-public interface UserRepository extends MongoRepository<Users, Integer> {
+public interface UserRepository extends MongoRepository<Users, Long> {
     Users findByUsername(String username);
 }
